requirements.txt
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
firebase-admin==4.1.0
Flask==1.1.2
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
httplib2==0.17.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lti==0.9.5
lxml==4.5.1
MarkupSafe==1.1.1
msgpack==1.0.0
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
PyMySQL==0.9.3
python-firebase==1.2
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
